Fungus dialog integration (#10)
* Adding Fungus

* Starting to work on the flow chart - still need to double check if it correctly connects

* Fixing a minor error found in code for game manager

* Moving around some items and having cards go in order of the Fungus flow chart

* Updating code so that the card reads from the block dialogue in the Fungus flow chart

* Temp push or reverting all previous changes

* Adding the Fungus Editor Resource assets

* Updates based on updated main - still testing out the flowchart functions

* Fungus Recognizes Swiping

Added "Day 1" Folder and Scene
When game is played, user is prompted to swipe left or right
When user swipes left or right, user will be prompted with the direction they chose to swipe.

* Flowchart Working but Slow

Flowchart is working but the timing for the choices is spread out so the player needs to wait a few seconds before the card is ready to swipe again. This will need to be addressed before pulling to main.

* Fungus Logic Working More Smooth

Game Manager handles most of the work. Fungus Logic is working. Only issue is the cards themselves shake.

* Removed Extra Fungus Examples

* Fixed Build Settings

Needed to enable a scene in the build settings

* mild fixes to branch

Merged main to branch and made needed fixes

* moved cardsdata

moved cards data to cards folder

---------

Co-authored-by: Kelli Porter <kelli.porter1998@gmail.com>
Co-authored-by: Ben Armour <Armourbl@mail.uc.edu>