Updated Day 1 and Day 2A
Issue: When navigating through the story, status variables would not update correctly.
Solution: Simplified GameManager.cs, Added all three variable types to blocks that cause changes. Variables that are not changing will be set to 0. Additionally, adding all three variables to the block that follows and setting them to 0 stops the variables from changing when they should not.

Day 1 and Day 2A are completed. Working on Day 2B and 2C next.